organized some src files
