+++ CMakeLists.txt
@@ -11,19 +11,19 @@ include_directories(${SDL2_INCLUDE_DIRS} ${SDL2_IMAGE_INCLUDE_DIRS})
 add_executable(sdlGameEngine
 src/camera/camera.cpp
 src/camera/camera.h
-src/character.cpp
-src/character_factory.cpp
-src/character_factory.h
-src/character.h
+src/character/character.cpp
+src/character/character_factory.cpp
+src/character/character_factory.h
+src/character/character.h
+src/character/npc.cpp
+src/character/npc.h
+src/character/player.cpp
+src/character/player.h
 src/game.cpp
 src/game.h
 src/logger.cpp
 src/logger.h
 src/main.cpp
-src/npc.cpp
-src/npc.h
-src/player.cpp
-src/player.h
 src/tile/dirt_tile.cpp
 src/tile/dirt_tile.h
 src/tile/tile.cpp
@@ -32,8 +32,8 @@ src/tile/tile_factory.h
 src/tile/tile.h
 src/tile/water_tile.cpp
 src/tile/water_tile.h
-src/tile_map.cpp
-src/tile_map.h
+src/tile/tile_map.cpp
+src/tile/tile_map.h
 src/button.h
 src/button.cpp
 src/main_menu.h
