+++ src/character/character.h
