+++ src/character/character_factory.cpp
