+++ src/character/character_factory.h
