+++ src/character/npc.cpp
