+++ src/character/npc.h
