+++ src/game.cpp
@@ -6,7 +6,6 @@
 
 #include "game.h"
 #include "camera/camera.h"
-#include "tile_map.h"
 #include "logger.h"
 #include "menu.h"
